+++ src/config/merge-configs.test.ts
@@ -130,8 +130,9 @@ test("Colliding bundle with invalid collision rule on incoming bundle", t => {
 	t.throws(
         () => MergeConfig([config1, config2]),
         {
-            instanceOf: Error,
-            message: "Exception raised while merging bundle 'testBundle' in the raw configuration at index '1'.",
+            instanceOf: RangeError,
+            message: "Exception raised while merging bundle 'testBundle' in the raw configuration at index '1'. \n"
+                + "Unexpected input 'badCollisionHandler' for 'onCollision' option of next bundle.",
         }
     );
 });